Broken down CategoryFilter componenet into smaller ones
