+++ client/components/index.js
@@ -8,9 +8,6 @@ export {default as UserHome} from './user/user-home'
 export {Login, Signup} from './auth/auth-form'
 export {default as AddressBook} from './sidePanel/addressBook/AddressBook'
 export {
-  default as SelectedCategories
-} from './sidePanel/categoryFilter/SelectedCategories'
-export {
   default as CategoryFilter
 } from './sidePanel/categoryFilter/CategoryFilter'
 export {default as Map} from './map/Map'
